+++ python/cudf/cudf/core/column/column.py
@@ -1037,7 +1037,6 @@ class ColumnBase(Serializable, BinaryOperand, Reducible):
         replacement: ColumnBase | list,
         *,
         null_cast_dtype: DtypeObj | None = None,
-        strict_type: bool = True,
     ) -> tuple[ColumnBase, ColumnBase]:
         to_replace_col = as_column(to_replace)
         if null_cast_dtype is not None and to_replace_col.is_all_null:
@@ -1047,14 +1046,9 @@ class ColumnBase(Serializable, BinaryOperand, Reducible):
         if null_cast_dtype is not None and replacement_col.is_all_null:
             replacement_col = replacement_col.astype(null_cast_dtype)
 
-        if strict_type:
-            type_matches = type(to_replace_col) is type(replacement_col)
-        else:
-            type_matches = isinstance(to_replace_col, type(replacement_col))
-
-        if not type_matches:
+        if type(to_replace_col) is not type(replacement_col):
             raise TypeError(
-                f"to_replace and value should be of same types,"
+                "to_replace and value should be of same types,"
                 f"got to_replace dtype: {to_replace_col.dtype} and "
                 f"value dtype: {replacement_col.dtype}"
             )
@@ -1066,16 +1060,20 @@ class ColumnBase(Serializable, BinaryOperand, Reducible):
         old_col: ColumnBase,
         new_col: ColumnBase,
     ) -> tuple[plc.Column, plc.Column]:
-        with old_col.access(mode="read", scope="internal"):
-            with new_col.access(mode="read", scope="internal"):
-                columns = plc.stream_compaction.stable_distinct(
-                    plc.Table([old_col.plc_column, new_col.plc_column]),
-                    keys=[0],
-                    keep=plc.stream_compaction.DuplicateKeepOption.KEEP_LAST,
-                    nulls_equal=plc.types.NullEquality.EQUAL,
-                    nans_equal=plc.types.NanEquality.ALL_EQUAL,
-                ).columns()
-        return cast(tuple[plc.Column, plc.Column], tuple(columns))
+        """Deduplicate old/new mapping with keep-last semantics.
+
+        This replicates pandas' behavior when to_replace has duplicates:
+        pandas processes replacements sequentially, so the last occurrence wins.
+        """
+        with access_columns(old_col, new_col, mode="read", scope="internal"):
+            columns = plc.stream_compaction.stable_distinct(
+                plc.Table([old_col.plc_column, new_col.plc_column]),
+                keys=[0],
+                keep=plc.stream_compaction.DuplicateKeepOption.KEEP_LAST,
+                nulls_equal=plc.types.NullEquality.EQUAL,
+                nans_equal=plc.types.NanEquality.ALL_EQUAL,
+            ).columns()
+        return columns[0], columns[1]
 
     def _find_and_replace_with_dedup(
         self,
@@ -1083,6 +1081,7 @@ class ColumnBase(Serializable, BinaryOperand, Reducible):
         new_col: ColumnBase,
         result_dtype: DtypeObj,
     ) -> ColumnBase:
+        """Apply find/replace using keep-last mapping and null handling."""
         old_plc, new_plc = ColumnBase._dedupe_find_and_replace_mapping(
             old_col, new_col
         )
@@ -3011,38 +3010,6 @@ def check_invalid_array(shape: tuple, dtype: np.dtype) -> None:
         raise ValueError("Data must be 1-dimensional")
     elif dtype == "float16":
         raise TypeError("Unsupported type float16")
-
-
-class UnderlyingDtypeReplaceMixin(ColumnBase):
-    _UNDERLYING_DTYPE: np.dtype
-
-    def find_and_replace(
-        self,
-        to_replace: ColumnBase | list,
-        replacement: ColumnBase | list,
-        all_nan: bool = False,
-    ) -> Self:
-        if not isinstance(to_replace, type(self)):
-            to_replace = as_column(to_replace)
-            if to_replace.can_cast_safely(self.dtype):
-                to_replace = to_replace.astype(self.dtype)
-        if not isinstance(replacement, type(self)):
-            replacement = as_column(replacement)
-            if replacement.can_cast_safely(self.dtype):
-                replacement = replacement.astype(self.dtype)
-        if isinstance(to_replace, type(self)):
-            to_replace = to_replace.astype(self._UNDERLYING_DTYPE)
-        if isinstance(replacement, type(self)):
-            replacement = replacement.astype(self._UNDERLYING_DTYPE)
-        try:
-            result = (
-                self.astype(self._UNDERLYING_DTYPE)
-                .find_and_replace(to_replace, replacement, all_nan)
-                .astype(self.dtype)
-            )
-        except TypeError:
-            return cast("Self", self.copy(deep=True))
-        return cast("Self", result)
 
 
 def as_column(
